🐞 fix: 분양글 상세조회 통계 테이블 타입 생성자 수정

+++ src/main/java/com/adoptpet/server/adopt/dto/response/AdoptDetailResponseDto.java
@@ -80,8 +80,8 @@ public class AdoptDetailResponseDto {
     @AllArgsConstructor
     public static class Context {
         private String context;
-        private long bookmark;
-        private long chat;
+        private Integer bookmark;
+        private Integer chat;
     }
 
     @Getter
